backend/requirements.txt
# Core Web Framework
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
flask-migrate==4.0.5  # For database migrations

# API and HTTP
requests==2.31.0
marshmallow==3.20.1  # Data serialization/validation

# Environment and Configuration
python-dotenv==1.0.0

# Natural Language Processing
nltk==3.8.1
pandas==2.2.0

# Optional NLP (Windows-friendly alternatives)
# Comment out or remove if installation fails
textblob==0.17.1  # Simple NLP
spacy-legacy==3.0.12  # Lighter NLP package

# Background Tasks
celery==5.3.6
redis==5.0.1  # Celery broker and result backend

# Development and Testing
pytest==7.4.4
pytest-flask==1.3.0
coverage==7.4.0

# Logging and Monitoring
structlog==24.1.0

# Security
flask-talisman==1.1.0
bcrypt==4.1.2

# API Documentation
flask-restx==1.3.0

# Performance and Caching
flask-caching==2.1.0

# Deployment
gunicorn==21.2.0

# Code Quality and Type Checking
black==24.1.1
mypy==1.8.0

# Additional Utilities
python-dateutil==2.9.0
arrow==1.3.0

# Optional: Alternative ML libraries (more Windows-friendly)
scikit-learn==1.3.2  # Machine learning
gensim==4.3.2  # Topic modeling and NLP